feat(app): minor styling and bug fixes
in the drawer component we removed the base height because we determined that the input button should always be at the bottom of the screen for mobile devices, this may change later. in the root page we added some comments and a class to keep the input box at the bottom of the screen, we also added the onfocus event to the input box to attempt to scroll the input element into view on mobile devices, this sometimes works, i'm looking for a better solution. in the index.ts file i added some test commands to test how the buttons work in the new layout. changed the name of the reminder command to have captial letters and spaces. in the revised command selector the search algorithm was modified to ignore spaces and case, and use a positional search instead of a substring search, so the queried command must start with the same string as the input, not an arbitrary position in the string, the main effect this has is omitting a lot of unnecessary matches. in the same file, i cleaned up the classes to appear a bit nicer in their parent, even if they're only here temporarily.