feat(app): mostly working on creating modular user input
added a few ideas to add for next time. moved choice, date, and text arg input components into a folder called "args" and propogated changes in dependencies. added the input selector which is a more flexible and modular version of the command selector which can take any type of input object instead of just commands, this is somewhat redundant in the scope of commands but will allow for much easier additions when more features are added. added the selectable option generic type and AnySelectableOption union type which help facilitate the input selectors dynamic behavior. renamed the command option interface to command argument for better clarity, this may become a type in the future. removed the two unused console logs in the root page component.