feat(app): refactored command selector for generic ux
completely switched from command selector to input selector for the generic capability and better design. added a wip scroll on focus function to scroll the ui when an input element is focused. added command context option within selectable option to give context as an object to allow arguments to reference their parent. moved the input element of the drawer into the input selector so it can be displayed an hidden as needed without passing values around. removed a few test instances of create reminder in the index cause like 20 is unnecessary. refactored the input selector to properly utilize the command context and added the ability to add and remove arguments, when adding this change a bug arose that prevented the deletion of multiple arguments because the current selectable option was used in the setOptionValue function, this was removed in exchange for a function argument to directly target a specific option instead of the active option. the main instance of this occuring is when the current option is a command and the user clicks the trash button to clear the existing value of an argument and there were multiple set argument values, previously when this occured the function would set the current option to the cleared argument and it would softlock the user in the drawer, with the function argument and the command context this bug is fixed. some additional text and buttons that utilize the context have also been implemented. there's a work in progress submit button to run the commmand but this is broken because of both a type issue and a state issue, it's the next thing to be addressed.