feat(app): small logic refactor for mvp
the main purpose of this update was to push towards getting an mvp working so some temporary fixes were added and will be made more reliable later. the reminder service was added to the input selector component as a probs for the context, this will be wrapped in a service context object later. the inputSelector selectable option title markdown was refactored to use an inline ternary expression to shorten the logic, this same approach was also applied to the set command argument value buttons where only the displayed text is different. the run command button was refactored to use the command context object instead of the current selectable option, I need to do this to the rest of this component so that behavior is consistent with the state's design. i also added arguments and passing the service context to the execute function so that function works as expected, this will probably need to be reworked, mainly with how values are passed to the execute function and how they're managed so passing them is easier and more implicit. I also added a couple of better comments in some spots.